feat: make cutoff points configurable
